// src/utils/enums/enums.ts
export enum ErrorCode {
    INTERNAL_ERROR = "INTERNAL_ERROR",
    VALIDATION_FAILED = "VALIDATION_FAILED",
    REQUIRED_TOKEN = "REQUIRED_TOKEN",
    EXPIRED_TOKEN = "EXPIRED_TOKEN",
    INVALID_TOKEN_TYPE = "INVALID_TOKEN_TYPE",
    INVALID_TOKEN_CLAIM = "INVALID_TOKEN_CLAIM",
    INVALID_ROLES_CLAIM = "INVALID_ROLES_CLAIM",
    INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
    INVALID_REQUEST = "INVALID_REQUEST",
    NOT_FOUND_USER = "NOT_FOUND_USER",
    DUPLICATE_USER = "DUPLICATE_USER",
    DELETED_USER = "DELETED_USER",
    MISSING_REQUIRED_DATA = "MISSING_REQUIRED_DATA",

    CANNOT_DELETE_ADMIN = "CANNOT_DELETE_ADMIN",

    // Role Code
    ROLE_NOT_FOUND = "ROLE_NOT_FOUND",
    DUPLICATE_ROLE = "DUPLICATE_ROLE",
    ROLE_DELETED = "ROLE_DELETED",
    ROLE_IS_IN_USE = "ROLE_IS_IN_USE",

    ZONE_NOT_FOUND = "ZONE_NOT_FOUND",
    DUPLICATE_ZONE = "DUPLICATE_ZONE",
    ZONE_IS_IN_USE = "ZONE_IS_IN_USE",

    AREA_NOT_FOUND = "AREA_NOT_FOUND",
    AREA_IS_IN_USE = "AREA_IS_IN_USE",

    BUILDING_IS_IN_USE = "BUILDING_IS_IN_USE",
    BUILDING_NOT_FOUND = "BUILDING_NOT_FOUND",

    APARTMENT_NOT_FOUND = "APARTMENT_NOT_FOUND",
    APARTMENT_IS_RENTED = "APARTMENT_IS_RENTED",
}

export enum SuccessCode {
    OPERATION_SUCCESS = "OPERATION_SUCCESS"
}